integrated auth with appropriate routes

+++ routes/awards.js
@@ -1,7 +1,8 @@
 const express = require('express');
 var router = express.Router();
 const fs = require('fs');
-const db = require ('../db.js'); 
+const db = require ('../db.js');
+const auth = require('../utilities/authenticate');
 
 // init table
 db.query("CREATE TABLE IF NOT EXISTS emp_award (award_id INTEGER PRIMARY KEY AUTO_INCREMENT, award_type INTEGER, awardee_name TEXT, awardee_dept TEXT, awardee_region TEXT, awardee_email TEXT, awarder_ID INTEGER, timestamp INTEGER, FOREIGN KEY (awarder_ID) REFERENCES emp_user(user_id))", [], (err) => {
@@ -12,6 +13,9 @@ db.query("CREATE TABLE IF NOT EXISTS emp_award (award_id INTEGER PRIMARY KEY AUT
 
 // routes
 router.get('/search/:field/:value', (req, res) => {
+    if( auth.isLoggedIn() === 0 ){
+        return;
+    }
     if( req.params['field'] === 'id' || req.params['field'] === 'award_type' || req.params['field'] === 'awardee_email' || req.params['field'] === 'awardee_name' || req.params['field'] === 'awarder_ID' || req.params['field'] === 'timestamp'){
         var parameter = req.params['field'];
     } else {
@@ -25,6 +29,9 @@ router.get('/search/:field/:value', (req, res) => {
 });
 
 router.post('/addAward', (req, res) => {
+    if( auth.isLoggedIn() === 0 ){
+        return;
+    }
     if( req.body['award_type'] === undefined || req.body['awardee_name'] === undefined || req.body['awardee_email'] == undefined || req.body['awarder_ID'] === undefined || req.body['timestamp'] === undefined ){
         res.sendStatus(400);
         return;
@@ -38,6 +45,9 @@ router.post('/addAward', (req, res) => {
 });
 
 router.post('/deleteAward/', (req, res) => {
+    if( auth.isLoggedIn() === 0 ){
+        return;
+    }
     db.query("DELETE FROM awards where ? = ?", [req.body['field'], req.body['value']], (err) => {
         if(err){
             console.error(err);
